added new functionalities

+++ applications/seat/seat.go
@@ -3,8 +3,8 @@ package seat
 type Seat struct {
 	SeatID    string  `json:"seatID" validate:"required"`
 	SeatType  string  `json:"seatType" validate:"requried"`
-	PriceGel  float32 `json:"priceGel" validate:"requried"`
-	PriceInr  float32 `json:"priceInr" validate:"requried"`
+	PriceGel  float64 `json:"priceGel" validate:"requried"`
+	PriceInr  float64 `json:"priceInr" validate:"requried"`
 	Available int     `json:"available"`
 	Notes     string  `json:"notes,omitempty"`
 }
